requirements.txt
pyyaml
langdetect
openai
dashscope
rich
pytest
